Organized Scripts into folders
