+++ DebuggerGame/Assets/Scripts/Action Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d6d5078ea8e0194eb26e39232f5a075
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
