+++ DebuggerGame/Assets/Scripts/BoardObject Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2bccee5d3acf9214890e58036a336797
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
