+++ DebuggerGame/Assets/Scripts/BoardObject Scripts/Arthropod Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b9f6afa5d7229db46a58e5407451a93b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
