+++ DebuggerGame/Assets/Scripts/BoardObject Scripts/Arthropod Scripts/Arthropod.cs.meta
