+++ DebuggerGame/Assets/Scripts/BoardObject Scripts/Arthropod Scripts/StationaryArthropod.cs.meta
