+++ DebuggerGame/Assets/Scripts/BoardObject Scripts/Player.cs.meta
