+++ DebuggerGame/Assets/Scripts/Inventory Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 824f35812f6e37b49b48d0e1865fa69c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
