+++ DebuggerGame/Assets/Scripts/Inventory Scripts/ArthropodData.cs.meta
