+++ DebuggerGame/Assets/Scripts/Inventory Scripts/ArthropodDatabase.cs
