+++ DebuggerGame/Assets/Scripts/UI Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 55ee86e813c07a646aae394347979d52
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
